requirements.txt
kafka-python
numpy
opencv-python
torch
pandas
onnx
grpcio
protobuf
fastapi
uvicorn
aiohttp
kubernetes
prometheus-client
python-multipart
python-dotenv
